Fix issues with calling Polymer.Debouncer
On Google hosted Gerrit the image-diff plugin is broken. The error in
the js console is:

`Polymer.Debouncer.debounce is not a function`

This has to do with the closure compilation, where the function ends up
being renamed to

`Debouncer$$module$third_party$javascript$polymer$v2$polymer$lib$utils$debounce$debounce`

This plugin is in maintenance mode, so we are looking for a quick fix.
Apparently the debouncing is not really required and can simply be
removed. I have tested this with both Bazel and Blaze builds and the
Dev Helper.

Google-Bug-Id: b/217942160
Release-Notes: skip
Change-Id: Ie183a3c1f68e9a2c2c1cd3a374d4d0231393afa0